remote.c: remove autotoolized conditional includes
these are the only headers that are conditionally included depending
on configure checks in the entire source tree, so we remove it for
consistency and just include them unconditionally.
this saves us from writing configure checks and implicit declaration
of inet_ntoa if the right HAVE_ define is not passed.