remove unportable,glibc/gcc specific remote profiling support
the code is not 64 bit safe and depends on glibc specifics.
the purpose of the code was to enable GBA developers to profile
their GBA binaries. since probably less than a 1000th of every
VBA user needed that feature, it's quite safe to remove it...
those that still need it can use the win32 version, or the old code.