final cleanup to SDL1-compatible audio code
this makes GBA sound perfect (no buffer-underruns at all), and GB
sound almost - there seems to be an underrun every ~10 seconds,
apparently there's a small calculation error in the GBsound code,
not producing the required amount of samples for the elapsed time.
this is mitigated as good as possible, with almost no audible
disturbances.

the major difficulty here is that SDL uses the provided sample
value only as a hint, and then returns in the "obtained" parameter
of SDL_OpenAudio() the buffer size it wants to have in each
callback invocation, which is unlikely to be identical with the
requested values.

the initial plan was to modify the code in the GB/GBA audio
producers to work with a customizable output buffer size and then
adapt the main audio buffer to the size SDL requests, but it turned
out that the output buffer and its size is hardwired into too many
places; therefore the frontend has to use its own buffering and
receive/send data in differently sized chunks.

the parent commit of this commit, which i sourced from another VBA
fork didn't actually improve things, supposedly its main purpose
was to fix the freeze on shutdown which i already fixed in
64d6610adeb4aba35fe2a7defc87aee85d3f49b1 .

that commit, bf7904f634bcd52d7722f9001283d0533ca2969d ,
caused the sound to stutter whenever a buffer underrun happened,
based the values it used on guesswork rather than science, and
invoked UB by using memcpy on overlapping regions.
however, it got rid of the overcomplicated junk code by
"deathpudding" which turned the audio code in an incomprehensible
mess, so i used it as a basis for my cleanup.