fix "white screen" issue with pokemon games
pokemon games ship with 1MBit flash and during startup do some
arithmetic on the values of flashManufacturer etc that are set
in flashSetSize(0x20000). if the values don't match the
expectations, the code just goes into an endless loop and never
paints anything onto the screen.
i spent hours looking for a generic solution to this problem
(like reading some bit in the romheader that says "using 1mbit")
but i couldn't find anything usable. therefore i adopted the
approach of another GBA emulator, namely checking the 4 letter
gameid against a list containing ids of known roms, and set
the romsize to 1mbit if it wasn't set already.